refactor(connection): rename connection
Rename connection to connectionController as it handles logic.
